Updated graphic interface.
Added a header label for the frame.
Added a static text field for the IP address of the server.
Added a dynamic text field for the port number of the server.
Added a check box option for deciding whether user should be kicked off
upon hitting a mine.
Added radio buttons for the two options for making the Minesweeper
board.